validation added

+++ proyecto-empleados/public/index.php
@@ -39,6 +39,11 @@ function getConnection() {
     $dbh->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
     return $dbh;
 }
+
+
+
+
+
 //Obtener listado personas
 function ObtenerPersonas($response) {
     $sql = "SELECT nationalId, name, lastname, age, originPlanet, pictureUrl FROM people";
@@ -46,7 +51,7 @@ function ObtenerPersonas($response) {
         $stmt = getConnection()->query($sql);
         $personas = $stmt->fetchAll(PDO::FETCH_OBJ);
         $db = null;
-        
+
         return json_encode($personas);
     } catch(PDOException $e) {
         echo '{"error":{"text":'. $e->getMessage() .'}}';
@@ -60,16 +65,18 @@ function ObtenerPersona($request) {
         $stmt = getConnection()->query($sql);
         $personas = $stmt->fetchAll(PDO::FETCH_OBJ);
         $db = null;
-	if ($personas == null) { return $response->withStatus(404);  }
+	if ($personas == null) {return header("Status: 404 Not Found"); }
         return json_encode($personas);
     } catch(PDOException $e) {
-                return $response->withStatus(404)
-            ->withHeader('Content-Type', 'text/html')
-            ->write('Something went wrong!');
+                return header("Status: 500 Error");
     }
 }
 //Agregar Persona
 function AgregarPersona($request) {
+    if ($request->getContentType() !== "application/json"){
+        return header("Status: 400 Content-Type should be application/json");
+    }
+    else{
     $emp = json_decode($request->getBody());
     $sql = "INSERT INTO people (nationalId, name, lastname, age, originPlanet, pictureUrl) VALUES (:nationalId,:name,:lastname,:age,:originPlanet,:pictureUrl)";
 
@@ -87,40 +94,69 @@ function AgregarPersona($request) {
         echo json_encode($emp);
     } catch(PDOException $e) {
         echo '{"error":{"text":'. $e->getMessage() .'}}';
+        header("Status: 500 Error");
+    }
     }
 }
 
 //Actualizar Persona
-function ActualizarPersona($request) {
-    $emp = json_decode($request->getBody());
-    $sql = "UPDATE people SET nationalId = :nationalId, name = :name, lastname = :lastname, age = :age, originPlanet = :originPlanet, pictureUrl = :pictureUrl where nationalId = :nationalId";
+function ActualizarPersona($request)
+{
+    if ($request->getContentType() !== "application/json") {
+        return header("Status: 400 Content-Type should be application/json");
+    } else {
+        $emp = json_decode($request->getBody());
 
-    try {
-        $db = getConnection();
-        $stmt = $db->prepare($sql);
-        $stmt->bindParam("nationalId", $emp->nationalId);
-        $stmt->bindParam("name", $emp->name);
-        $stmt->bindParam("lastname", $emp->lastname);
-        $stmt->bindParam("age", $emp->age);
-        $stmt->bindParam("originPlanet", $emp->originPlanet);
-        $stmt->bindParam("pictureUrl", $emp->pictureUrl);
-        $stmt->execute();
+        $nationalId = $request->getAttribute('nationalId');
+        $sql = "SELECT nationalId, name, lastname, age, originPlanet, pictureUrl FROM people where nationalId = '$nationalId';";
+        $stmt = getConnection()->query($sql);
+        $personas = $stmt->fetchAll(PDO::FETCH_OBJ);
+											 
+													 
+										   
+															 
+														 
+						 
         $db = null;
-        echo json_encode($emp);
-    } catch(PDOException $e) {
-        echo '{"error":{"text":'. $e->getMessage() .'}}';
+        if ($personas == null) {return header("Status: 404 People Not Found"); }
+
+
+
+        $sql = "UPDATE people SET nationalId = :nationalId, name = :name, lastname = :lastname, age = :age, originPlanet = :originPlanet, pictureUrl = :pictureUrl where nationalId = :nationalId";
+
+        try {
+            $db = getConnection();
+            $stmt = $db->prepare($sql);
+            $stmt->bindParam("nationalId", $emp->nationalId);
+            $stmt->bindParam("name", $emp->name);
+            $stmt->bindParam("lastname", $emp->lastname);
+            $stmt->bindParam("age", $emp->age);
+            $stmt->bindParam("originPlanet", $emp->originPlanet);
+            $stmt->bindParam("pictureUrl", $emp->pictureUrl);
+            $stmt->execute();
+            $db = null;
+            echo json_encode($emp);
+        } catch (PDOException $e) {
+            echo '{"error":{"text":' . $e->getMessage() . '}}';
+            header("Status: 500 Error");
+        }
     }
 }
 //Eliminar 1 persona
 function EliminarPersona($request) {
           $nationalId = $request->getAttribute('nationalId');
+    $sql = "SELECT nationalId, name, lastname, age, originPlanet, pictureUrl FROM people where nationalId = '$nationalId';";
+    $stmt = getConnection()->query($sql);
+    $personas = $stmt->fetchAll(PDO::FETCH_OBJ);
+    $db = null;
+    if ($personas == null) {return header("Status: 404 People Not Found"); }
       $sql = "DELETE FROM people where nationalId = '$nationalId'";
-  echo $sql;
+
   try {
         $stmt = getConnection()->query($sql);
         $personas = $stmt->fetchAll(PDO::FETCH_OBJ);
         $db = null;
-        if ($personas == null) 
+        if ($personas == null)
 	{ return $response->withStatus(404);  } else {
         	return json_encode($personas); }
     }  catch(PDOException $e) {
